packages/cli/templates/plugins/ai-pipeline/requirements.txt
pyspark==3.5.1
fastapi==0.111.0
uvicorn[standard]==0.30.0
kafka-python-ng==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
pydantic==2.7.0
mlflow-skinny==2.13.0
clickhouse-connect==0.7.19
